Refactor: move test codes into /test/services
